refactor code and add support with other opcodes for statechange
